pyproject.toml
[project]
name = "google-skills"
version = "1.0.0"
description = "Agent Skills for Google AI services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.0",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
